ollama_rag/requirements.txt
beautifulsoup4==4.12.2
chromadb==0.4.20
langchain==0.0.350
python-dotenv==1.0.0
google-api-python-client
google-auth-httplib2 
google-auth-oauthlib
PyPDF2